Test Python 3.7 with a poetry that supports it
Poetry [dropped support for Python
3.7](https://github.com/python-poetry/poetry/pull/7674) in [version
1.6.0](https://python-poetry.org/history/#160---2023-08-20). There's no
reason why this plugin cannot work in a python 3.7 environment, provided
a poetry that supports that environment is installed. Since poetry 1.5.1
was the last version to support Python 3.7 this change switches the test
setup to use poetry 1.5.1 in the case of the Python 3.7 branch of the
testing matrix.